S11P12D109-151 fix: 공연 Repository 키 자료형 Long -> Integer 수정

+++ Colorpl-Backend/src/main/java/com/colorpl/show/domain/ShowRepository.java
@@ -2,5 +2,5 @@ package com.colorpl.show.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ShowRepository extends JpaRepository<Show, Long> {
+public interface ShowRepository extends JpaRepository<Show, Integer> {
 }
